+++ src/main/kotlin/tech/gdragon/api/pawa/Pawa.kt
@@ -8,6 +8,7 @@ import tech.gdragon.api.commands.RecoverResult
 import tech.gdragon.api.commands.safeFile
 import tech.gdragon.api.tape.extractDuration
 import tech.gdragon.api.tape.queueFileIntoMp3
+import tech.gdragon.api.tape.remuxWithXingHeader
 import tech.gdragon.data.Datastore
 import tech.gdragon.db.Database
 import tech.gdragon.db.dao.*
@@ -159,13 +160,16 @@ open class Pawa(val db: Database, val config: PawaConfig = PawaConfig.invoke()) 
     val mp3Exists: Boolean = when {
       mp3File.exists() -> {
         logger.info { "Recovering from mp3 file." }
+        remuxWithXingHeader(mp3File)
         true
       }
 
       // This is a side effect
       queueFile.exists() -> {
         logger.info { "Recovering $sessionId from queue file." }
-        queueFileIntoMp3(queueFile, mp3File).exists()
+        queueFileIntoMp3(queueFile, mp3File)
+        remuxWithXingHeader(mp3File)
+        mp3File.exists()
       }
 
       queueFile.exists().not() && mp3File.exists().not() -> {