+++ src/main/kotlin/tech/gdragon/listener/BaseAudioRecorder.kt
@@ -19,7 +19,7 @@ import tech.gdragon.BotUtils
 import tech.gdragon.api.pawa.Pawa
 import tech.gdragon.api.tape.addCommentToMp3
 import tech.gdragon.api.tape.queueFileIntoMp3
-import tech.gdragon.api.tape.writeVbrTag
+import tech.gdragon.api.tape.remuxWithXingHeader
 import tech.gdragon.data.Datastore
 import tech.gdragon.db.dao.Channel
 import tech.gdragon.db.dao.Guild
@@ -269,9 +269,7 @@ abstract class BaseAudioRecorder(
         val recordingFile = File("$dataDirectory/recordings/$ulid.$fileFormat")
         queueFileIntoMp3(queue, recordingFile)
 
-        if (vbr && fileFormat == "mp3") {
-          lameEncoder?.let { writeVbrTag(it, recordingFile) }
-        }
+        remuxWithXingHeader(recordingFile)
 
         addCommentToMp3(recordingFile, recording?.speakers?.joinToString(prefix = "Speakers: ") { it.name })